perf: 优化配置
