+++ apps/settings/api/__init__.py
@@ -1,5 +1,7 @@
-from .common import *
+from .settings import *
 from .ldap import *
 from .wecom import *
 from .dingtalk import *
 from .feishu import *
+from .public import *
+from .email import *
