+++ apps/settings/serializers/__init__.py
@@ -1,7 +1,13 @@
 # coding: utf-8
 #
 
+from .basic import *
+from .im import *
+from .auth import *
 from .email import *
-from .ldap import *
 from .public import *
 from .settings import *
+from .security import *
+from .terminal import *
+from .advanced import *
+
