+++ apps/settings/serializers/auth/__init__.py
@@ -1,0 +1,4 @@
+from .cas import *
+from .ldap import *
+from .oidc import *
+from .radius import *
